Init adding posts(services, controller, migration)

+++ TowNews.WebUI/Controllers/CategoryController.cs
@@ -1,17 +1,25 @@
-﻿using Microsoft.AspNetCore.Identity;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using TopNews.Core.DTOS.Category;
+using TopNews.Core.DTOS.Post;
 using TopNews.Core.Interfaces;
 using TopNews.Core.Services;
+using TopNews.Core.Validation.Category;
+using X.PagedList;
 
 namespace TopNews.WebUI.Controllers
 {
+    [Authorize]
     public class CategoryController : Controller
     {
         private readonly ICategoryService _categoryService;
-        public CategoryController(ICategoryService categoryService)
+        private readonly IPostService _postService;
+
+        public CategoryController(ICategoryService categoryService, IPostService postService)
         {
             _categoryService = categoryService;
+            _postService = postService;
         }
         public async Task<IActionResult> Index()
         {
@@ -26,16 +34,48 @@ namespace TopNews.WebUI.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> AddNew(CategoryDTO model)
         {
-            await _categoryService.Create(model);
-            return RedirectToAction("Index");
+            var validator = new CreateValidation();
+            var validationResult = await validator.ValidateAsync(model);
+            if (validationResult.IsValid)
+            {
+                var result = await _categoryService.GetByName(model);
+                if (!result.Success)
+                {
+                    ViewBag.AuthError = "Category exists.";
+                    return View(model);
+                }
+                await _categoryService.Create(model);
+                return RedirectToAction(nameof(Index));
+            }
+            ViewBag.AuthError = validationResult.Errors[0];
+            return View(model);
         }
         public async Task<IActionResult> Delete(int id)
         {
-            var result = await _categoryService.Get(id);
-            return View(result);
+            var categoryDto = await _categoryService.Get(id);
+
+            if (categoryDto == null)
+            {
+                ViewBag.AuthError = "Category not found.";
+                return View();
+            }
+
+            List<PostDTO> posts = await _postService.GetByCategory(id);
+            ViewBag.CategoryName = categoryDto.Name;
+            ViewBag.CategoryId = categoryDto.Id;
+
+            int pageSize = 20;
+            int pageNumber = 1;
+            return View("Delete", posts.ToPagedList(pageNumber, pageSize));
         }
         public async Task<IActionResult> DeleteCategory(int id)
         {
+            var categoryDto = await _categoryService.Get(id);
+            if (categoryDto == null)
+            {
+                ViewBag.AuthError = "Category not found.";
+                return View();
+            }
             await _categoryService.Delete(id);
             return RedirectToAction(nameof(Index));
         }
@@ -48,9 +88,15 @@ namespace TopNews.WebUI.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(CategoryDTO model)
         {
-            await _categoryService.Update(model);
-            return RedirectToAction("Index", "Dashboard");
-            //return View("Index");
+            var validator = new CreateValidation();
+            var validationResult = await validator.ValidateAsync(model);
+            if (validationResult.IsValid)
+            {
+                await _categoryService.Update(model);
+                return RedirectToAction(nameof(Index));
+            }
+            ViewBag.AuthError = validationResult.Errors[0];
+            return View(model);
         }
     }
 }
